# server/openapi.yaml
openapi: 3.1.0
...
info:
  title: ANZH Beauty API
  version: '1.0.0'
  description: |
    Бэкенд мини-приложения ANZH Cosmetology.

    **Личность.** Все запросы, кроме витрины и здоровья, требуют
    сессии. Сессия выдаётся в обмен на `initData` из Telegram, подпись
    которого проверяется на сервере (HMAC + срок жизни 1 час). Роль
    берётся с сервера по `telegram_id`, а не из запроса.

    **Деньги.** Целое в минорных единицах плюс код валюты: `45000` и
    `GEL` означают 450.00 лари. Telegram Stars — отдельное поле
    `stars_amount`, это не валюта ISO-4217.

    **Время.** Всё в UTC, ISO-8601. Часовой пояс салона —
    `Asia/Tbilisi` — параметр конфигурации, а не свойство данных.

    **Итог визита считает сервер.** Клиент присылает состав, а не сумму.
servers:
  - url: '{base}/api/v1'
    description: Базовый адрес берётся из переменной окружения клиента
    variables:
      base:
        default: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /bookings:
    post:
      tags: [Записи]
      summary: Оставить заявку
      description: |
        Создаёт заявку в статусе `pending`. Подтверждает её человек —
        заявки друг друга не блокируют.

        Мастер выбирается по тому, кто ведёт эти процедуры; если
        прислан `staff_id`, он проверяется на соответствие.

        Сумму, скидку и длительность из тела запроса сервер игнорирует
        и считает сам.

        Ограничение частоты: 10 заявок за 5 минут на человека.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slugs, starts_at]
              properties:
                slugs:
                  type: array
                  items: { type: string }
                  minItems: 1
                  maxItems: 6
                starts_at: { type: string, format: date-time }
                staff_id: { type: [integer, 'null'], format: int64 }
                note:
                  type: [string, 'null']
                  maxLength: 1000
                  description: |
                    Свободный текст клиентки. Может содержать сведения о
                    здоровье, поэтому не попадает ни в логи, ни в текст
                    уведомлений.
      responses:
        '201':
          description: Заявка создана (или возвращена по тому же ключу идемпотентности)
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking: { $ref: '#/components/schemas/Booking' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: '`bad_slugs` · `bad_time` · `past_time` · `staff_mismatch` · `no_staff` · `bad_idempotency_key`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/TooMany' }
    get:
      tags: [Записи]
      summary: Мои записи
      responses:
        '200':
          description: Список, новые сверху
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /bookings/declined:
    get:
      tags: [Записи]
      summary: Отклонённые заявки за две недели
      description: Клиентка должна видеть отказ и его причину, а не тишину.
      responses:
        '200':
          description: Список
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /bookings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Записи]
      summary: Одна запись
      description: |
        Право проверяется на конкретный объект: своя запись, своя как у
        мастера, либо владелица. Иначе `403`, даже если сессия есть.
      responses:
        '200':
          description: Запись
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking: { $ref: '#/components/schemas/Booking' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Записи]
      summary: Сменить статус
      description: |
        Разрешённые переходы:

        | из | мастер может | клиентка может |
        |---|---|---|
        | `pending` | `confirmed`, `declined` | `cancelled` |
        | `confirmed` | `done`, `no_show`, `cancelled` | `cancelled` |
        | остальные | — | — |

        Отклонение требует причины: отказ без объяснения клиентка
        видит как молчание.

        Подтверждение занимает слот. Конфликт ловит уникальный индекс
        в базе, а не проверка в коде.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/BookingStatus' }
                reason: { type: [string, 'null'], maxLength: 500 }
      responses:
        '200':
          description: Статус изменён
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking: { $ref: '#/components/schemas/Booking' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`slot_taken` — время занято · `bad_transition` · `stale` — статус изменили одновременно'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422':
          description: '`bad_status` · `reason_required`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /staff/bookings:
    get:
      tags: [Записи]
      summary: День мастера и очередь заявок
      description: Мастер видит свои записи. `scope=all` действует только для владелицы.
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/BookingStatus' } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: scope, in: query, schema: { type: string, enum: [mine, all] } }
      responses:
        '200':
          description: Список
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingList' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /staff/{staffId}/busy:
    get:
      tags: [Записи]
      summary: Занятые слоты мастера
      description: |
        Занятыми считаются только подтверждённые записи. Ответ не
        раскрывает, кто записан и на что.
      parameters:
        - { name: staffId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Времена начала занятых визитов
          content:
            application/json:
              schema:
                type: object
                properties:
                  busy:
                    type: array
                    items: { type: string, format: date-time }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          description: '`bad_staff` · `bad_range` · `range_too_long` (больше 90 дней)'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <токен>`. Полезная нагрузка ровно такая:
        `{ telegram_id, app_source: "beauty", role }`. Поле `app_source`
        сверяется при каждом запросе — токен другого приложения это не
        открывает.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string, format: uuid }
      description: |
        Повтор с тем же ключом возвращает тот же результат, а не
        создаёт второй объект. Сеть теряет ответ, человек жмёт кнопку
        второй раз — без ключа получилось бы две записи.
  responses:
    Unauthorized:
      description: Нет сессии
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: { code: no_session, message: Нужно войти заново } }
...
    Forbidden:
      description: Роли или прав на объект не хватает
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Объекта нет
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    TooMany:
      description: Слишком часто. Заголовок `Retry-After` — через сколько секунд можно снова
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      description: |
        Единая форма на все отказы. Ни трассировки стека, ни текста
        запроса к базе: по ним читается устройство сервера, а в тексте
        запроса могут оказаться данные из анкеты.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: slot_taken }
            message: { type: string, example: Время уже занято }
...
    BookingStatus:
      type: string
      enum: [pending, confirmed, declined, done, no_show, cancelled]
    Booking:
      type: object
      properties:
        id: { type: string, format: uuid }
        telegram_id: { type: integer, format: int64 }
        staff_id: { type: integer, format: int64 }
        staff_name: { type: [string, 'null'] }
        starts_at: { type: string, format: date-time }
        duration_min: { type: integer }
        status: { $ref: '#/components/schemas/BookingStatus' }
        total_minor: { type: integer, format: int64 }
        currency: { type: string }
        discount_percent: { type: integer }
        client_note: { type: [string, 'null'] }
        decline_reason: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        items:
          type: array
          description: Цены зафиксированы на момент записи — прайс меняется, история нет.
          items:
            type: object
            properties:
              slug: { type: string }
              title_ru: { type: string }
              title_en: { type: string }
              price_minor: { type: integer, format: int64 }
    BookingList:
      type: object
      properties:
        bookings:
          type: array
          items: { $ref: '#/components/schemas/Booking' }
...